requirements.txt
# Core dependencies
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
numpy>=1.21.0
pillow>=9.0.0
einops>=0.6.0
omegaconf>=2.3.0
lpips>=0.1.4
easydict>=1.9
tensorboard>=2.10.0
huggingface_hub>=0.16.0
iopath>=0.1.10
scipy>=1.9.0
av>=10.0.0
xformers==0.0.32.post2
timm==1.0.25
# Additional dependencies for data preprocessing
opencv-python>=4.13.0,<4.14.0
tqdm
pandas
